Remove delayed evaluation
After careful consideration, I don't believe it is possible to require
consuming services to handle their own ACL expansions. This is because:

* We can only apply ReadPermission permission after full expansion. If
  we pass a consuming service an unexpanded ACL we are allowing them to
  see detail of other permission grants. The service would also have to
  explicitly ignore these unnecessary ACEs.

* Passing a consuming service a fully-expanded ACL is the minimum
  information that service needs to do its job. Requiring the service to
  have permission to perform its own lookups appears more secure, but
  isn't; passing the full ACL gives the service exactly the information
  needed and no more, without any additional configuration.

Additionally, requiring every consuming service to perform its own ACL
expansion makes the job much more complicated. The primary reason for
this was to avoid bootstrap problems with MQTT change-notify, but if we
are now using direct WebSockets instead this is no longer a problem. In
fact, a single WS to the Auth service is now simpler than a web of WSs
to each individual service required for ACL expansion.