Migrate the config schemas to an app
We will still need a Special, because we have different validation
requirements. But this allows us to access the etag mechanism and will
make it easier to create categorise the applications.

We will need to consider whether we need to validate that an App-UUID is
a member of Application on every request. This will partly depend on how
performant the recursive membership query is. The new Auth service will
be making similar queries during permission expansion.